# Proposed patch for key indices