@mrbotch 2015-12-14
Thanks for that. I guess Titan isn't particularly flexible towards schema changes so we may as well require an application server restart to refresh the indices cache for now.